Styling

+++ src/containers/Viewer/ChartTypes/PieChart.tsx
@@ -16,7 +16,7 @@ class PieChartStrategy implements ChartStrategy {
             if(formattedData.length <= 0){
                 formattedData.push([key[0].value, value[0].value])
             }
-            formattedData.push([key[1].value, parseInt(value[1].value)])
+            formattedData.push([key[1].value, parseFloat(value[1].value)])
         }
         return formattedData;
     }
